refactor: merged update submission state commands
