--- Source/Application/Kysect.Shreks.Application.Abstractions/Submissions/Commands/InvalidateSubmission.cs
@@ -1,11 +1,0 @@
-using Kysect.Shreks.Application.Dto.Study;
-using MediatR;
-
-namespace Kysect.Shreks.Application.Abstractions.Submissions.Commands;
-
-public static class InvalidateSubmission
-{
-    public record struct Command(Guid UserId, Guid SubmissionId) : IRequest<Response>;
-
-    public record struct Response(SubmissionDto Submission);
-}
