+++ Source/Application/Kysect.Shreks.Application.Commands/Commands/DeactivateCommand.cs
@@ -4,6 +4,7 @@ using Kysect.Shreks.Application.Commands.Contexts;
 using Kysect.Shreks.Application.Commands.Processors;
 using Kysect.Shreks.Application.Commands.Result;
 using Kysect.Shreks.Application.Dto.Study;
+using Kysect.Shreks.Application.Dto.Users;
 
 namespace Kysect.Shreks.Application.Commands.Commands;
 
@@ -12,12 +13,17 @@ public class DeactivateCommand : IShreksCommand<SubmissionContext, SubmissionDto
 {
     public async Task<SubmissionDto> ExecuteAsync(SubmissionContext context, CancellationToken cancellationToken)
     {
-        var command = new DeactivateSubmission.Command(context.IssuerId, context.Submission.Id);
+        var command = new UpdateSubmissionState.Command(
+            context.IssuerId, context.Submission.Id, SubmissionStateDto.Inactive);
+
         var response = await context.Mediator.Send(command, cancellationToken);
 
         return response.Submission;
     }
 
-    public Task<TResult> AcceptAsync<TResult>(IShreksCommandVisitor<TResult> visitor) where TResult : IShreksCommandResult
-        => visitor.VisitAsync(this);
+    public Task<TResult> AcceptAsync<TResult>(IShreksCommandVisitor<TResult> visitor)
+        where TResult : IShreksCommandResult
+    {
+        return visitor.VisitAsync(this);
+    }
 }
