+++ Source/Application/Kysect.Shreks.Application.Dto/Users/SubmissionStateDto.cs
@@ -1,0 +1,9 @@
+namespace Kysect.Shreks.Application.Dto.Users;
+
+public enum SubmissionStateDto
+{
+    Active,
+    Inactive,
+    Invalid,
+    Completed,
+}
